Make the termio subsystem loadable on its own.
This patch allows the net.didierverna.clon.termio subsystem to be loaded on
its own. This is needed for Quicklisp to find it and properly reference it.
See clon issue #9 and quicklisp-controller issue #20.

By making the component type a string as opposed to a symbol, ASDF will
attempt to UIOP:SAFE-READ-FROM-STRING it after the :defsystem-depends-on
dependencies have been loaded. At that time, the SB-GROVEL package will be
known.

2022-06-27  Jake Connor <jconnor@dwavesys.com>

	* termio/net.didierverna.clon.termio.asd
	(:net.didierverna.clon.termio): Stringify the
	GROVEL-CONSTANTS-FILE component type so as to avoid direct
	reference to the potentially unknown SB-GROVEL package.